requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
gender_guesser==0.4.0
gunicorn==21.2.0
pandas==2.3.1
psycopg2-binary
python-dotenv==1.1.1
pytz==2024.2
Requests==2.32.4
SQLAlchemy==2.0.37
waitress==3.0.2
weasyprint==65.1
wbiztool_client==1.0.0
Werkzeug==3.1.3
WTForms==3.2.1
